feat: setup function using boot api and test code
