+++ node/.env.example
@@ -7,17 +7,21 @@ BOOT_NODE=
 LISTEN_PORT=
 PROVIDER_URL=
 CONFIG_URL=
-
-#not required if wallets table is not empty
-REPORTER_PK=
-
 SUBMISSION_PROXY_CONTRACT=
 DELEGATOR_URL=
 
-#required to run klaytn_helper test
-#TEST_FEE_PAYER_PK=
+# (optional) required if wallets table is empty
+REPORTER_PK=
 
+# (optional) required to run klaytn_helper test
+TEST_FEE_PAYER_PK=
+
+# (optional) required to set private network
 PRIVATE_NETWORK_SECRET=
 
+# (optional) defaults to 8089, 9999 if not set
 BOOT_API_PORT=
 BOOT_LISTEN_PORT=
+
+# (optional) required if initializing host with boot api
+BOOT_API_URL=
