+++ node/cmd/boot/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"context"
 
-	"bisonai.com/orakl/node/pkg/libp2p"
+	"bisonai.com/orakl/node/pkg/libp2p/setup"
 
 	"github.com/rs/zerolog/log"
 )
@@ -12,7 +12,7 @@ func main() {
 	ctx := context.Background()
 	seed := "orakl-bootnode-test"
 
-	bootnode, err := libp2p.SetBootNode(ctx, 10010, seed)
+	bootnode, err := setup.SetBootNode(ctx, 10010, seed)
 	if err != nil {
 		log.Fatal().Err(err).Msg("Failed to set bootnode")
 	}
