+++ node/pkg/aggregator/main_test.go
@@ -11,7 +11,7 @@ import (
 	"bisonai.com/orakl/node/pkg/admin/utils"
 	"bisonai.com/orakl/node/pkg/bus"
 	"bisonai.com/orakl/node/pkg/db"
-	"bisonai.com/orakl/node/pkg/libp2p"
+	libp2p_setup "bisonai.com/orakl/node/pkg/libp2p/setup"
 	"github.com/gofiber/fiber/v2"
 	"github.com/rs/zerolog"
 )
@@ -53,12 +53,12 @@ func setup(ctx context.Context) (func() error, *TestItems, error) {
 	}
 	testItems.admin = admin
 
-	h, err := libp2p.MakeHost(10001)
+	h, err := libp2p_setup.MakeHost(10001)
 	if err != nil {
 		return nil, nil, err
 	}
 
-	ps, err := libp2p.MakePubsub(ctx, h)
+	ps, err := libp2p_setup.MakePubsub(ctx, h)
 	if err != nil {
 		return nil, nil, err
 	}
