+++ node/pkg/reporter/main_test.go
@@ -10,7 +10,7 @@ import (
 	"bisonai.com/orakl/node/pkg/admin/utils"
 	"bisonai.com/orakl/node/pkg/bus"
 	"bisonai.com/orakl/node/pkg/db"
-	"bisonai.com/orakl/node/pkg/libp2p"
+	libp2p_setup "bisonai.com/orakl/node/pkg/libp2p/setup"
 	"github.com/gofiber/fiber/v2"
 	"github.com/rs/zerolog"
 )
@@ -54,7 +54,7 @@ func setup(ctx context.Context) (func() error, *TestItems, error) {
 
 	testItems.admin = admin
 
-	h, ps, err := libp2p.Setup(ctx, "", 10001)
+	h, ps, err := libp2p_setup.Setup(ctx, "", 10001)
 	if err != nil {
 		return nil, nil, err
 	}
