+++ node/script/test_connection/main.go
@@ -5,7 +5,7 @@ import (
 	"flag"
 	"time"
 
-	"bisonai.com/orakl/node/pkg/libp2p"
+	libp2p_setup "bisonai.com/orakl/node/pkg/libp2p/setup"
 	"github.com/rs/zerolog/log"
 )
 
@@ -23,7 +23,7 @@ func main() {
 	}
 
 	startTime := time.Now()
-	_, ps, err := libp2p.Setup(ctx, *bootnode, *port)
+	_, ps, err := libp2p_setup.Setup(ctx, *bootnode, *port)
 	if err != nil {
 		log.Fatal().Err(err).Msg("Failed to setup libp2p")
 	}
